orders

+++ laravelfyp2021/database/migrations/2021_11_22_061910_create_orders_table.php
@@ -16,10 +16,10 @@ class CreateOrdersTable extends Migration
         Schema::create('orders', function (Blueprint $table) {
             $table->id();
             $table->bigInteger('user_id')->unsigned();
-            $table->decimal('subtotal');
-            $table->decimal('discount')->default(0);
-            $table->decimal('tax');
-            $table->decimal('total');
+            $table->float('subtotal');
+            $table->float('discount')->default(0);
+            $table->float('tax');
+            $table->float('total');
             $table->string('firstname');
             $table->string('lastname');
             $table->string('mobile');
